+++ src/main/java/wooteco/subway/domain/Path.java
@@ -19,37 +19,11 @@ public class Path {
     }
 
     public static Path from(List<Section> sections, Station departure, Station arrival) {
-        final GraphPath<Station, DefaultWeightedEdge> path = generatePath(sections, departure, arrival);
-        final List<Station> stations = path.getVertexList();
-        final int distance = (int) path.getWeight();
+        ShortestPath shortestPath = ShortestPath.generate(sections, departure, arrival);
+        final List<Station> stations = shortestPath.getPath();
+        final int distance = shortestPath.getDistance();
         final Cost cost = Cost.from(distance);
         return new Path(stations, distance, cost);
-    }
-
-    private static GraphPath<Station, DefaultWeightedEdge> generatePath(List<Section> sections, Station departure,
-                                                                 Station arrival) {
-        final WeightedMultigraph<Station, DefaultWeightedEdge> graph = new WeightedMultigraph<>(
-                DefaultWeightedEdge.class);
-
-        addSectionsToGraph(sections, graph);
-
-        final DijkstraShortestPath<Station, DefaultWeightedEdge> shortestPath = new DijkstraShortestPath<>(graph);
-        try {
-            return shortestPath.getPath(departure, arrival);
-        } catch (IllegalArgumentException e) {
-            throw new IllegalArgumentException("찾으시는 경로가 존재하지 않습니다!");
-        }
-    }
-
-    private static void addSectionsToGraph(List<Section> sections, WeightedMultigraph<Station, DefaultWeightedEdge> graph) {
-        for (Section section : sections) {
-            final Station upStation = section.getUpStation();
-            final Station downStation = section.getDownStation();
-
-            graph.addVertex(upStation);
-            graph.addVertex(downStation);
-            graph.setEdgeWeight(graph.addEdge(upStation, downStation), section.getDistance());
-        }
     }
 
     public List<Station> getStations() {
